Cleaned up exports from notebookPicker
